requirements.txt
langchain
langchain-core
langchain-openai
openai
langchain-anthropic
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
langchain-google-genai
transformers
pandas

